move creation of object to tests directory
