+++ tests/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright (c) 2020-2022, NVIDIA CORPORATION.
+  Copyright (c) 2020-2023, NVIDIA CORPORATION.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -343,6 +343,78 @@
             </dependencies>
         </profile>
         <profile>
+            <id>release311</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>311</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release312</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>312</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release312db</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>312db</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/312+-db/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release313</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>313</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release314</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>314</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release320</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>320</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/320+-noncdh-nondb/scala,${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
             <id>release321</id>
             <activation>
                 <property>
@@ -352,6 +424,18 @@
             </activation>
             <properties>
                 <spark.shim.test.sources>${project.basedir}/src/test/320+-noncdh-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release321db</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>321db</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/312+-db/scala</spark.shim.test.sources>
             </properties>
         </profile>
         <profile>
@@ -387,7 +471,19 @@
                 </property>
             </activation>
             <properties>
-                <spark.shim.test.sources>${project.basedir}/src/test/330+/scala,${project.basedir}/src/test/320+-noncdh-nondb/scala</spark.shim.test.sources>
+                <spark.shim.test.sources>${project.basedir}/src/test/330+/scala,${project.basedir}/src/test/320+-noncdh-nondb/scala,${project.basedir}/src/test/311+-nondb/scala</spark.shim.test.sources>
+            </properties>
+        </profile>
+        <profile>
+            <id>release330db</id>
+            <activation>
+                <property>
+                    <name>buildver</name>
+                    <value>330db</value>
+                </property>
+            </activation>
+            <properties>
+                <spark.shim.test.sources>${project.basedir}/src/test/312+-db/scala</spark.shim.test.sources>
             </properties>
         </profile>
         <profile>
