+++ tests/src/test/scala/com/nvidia/spark/rapids/DistributionSuite.scala
@@ -16,7 +16,7 @@
 
 package com.nvidia.spark.rapids
 
-import com.nvidia.spark.rapids.shims.{CustomBroadcastMode, ShimSparkPlan}
+import com.nvidia.spark.rapids.shims.ShimSparkPlan
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
